requirements.txt
lightgbm
scikit-learn
pandas
numpy
matplotlib
seaborn
shap
joblib
pytest
plotly
umap-learn